requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.8.0
django-filter==25.1
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
model-bakery==1.20.5
msgpack==1.1.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
whitenoise==6.11.0
zope.interface==8.0.1
